requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
pyyaml==6.0.2
